// src/app/enums/categorias.ts
export enum Categorias {
    ficcao = 'Ficção',
    biografia = 'Biografia',
    folclore = 'Folclore',
    literaturaBrasileira = 'Literatura Brasileira',
    colecoes = 'Coleções',
    genealogia = 'Genealogia',
    literaturaEstrangeira = 'Literatura Estrangeira',
    comportamento = 'Comportamento',
    humor = 'Humor',
    livrosRaros = 'Livros Raros'

}